change prediction

+++ tf_utils.py
@@ -26,9 +26,9 @@ def norm_probability(mu, sigma, sample):
 #Used to initialize weights for policy and value output layers
 def normalized_columns_initializer(std=1.0):
     def _initializer(shape, dtype=None, partition_info=None):
-        out = np.random.randn(*shape).astype(np.float32)
+        out = np.random.randn(*shape)
         out *= std / np.sqrt(np.square(out).sum(axis=0, keepdims=True))
-        return tf.constant(out)
+        return tf.constant(out, dtype=tf.float32)
     return _initializer
